+++ src/adblock.js
@@ -62,8 +62,8 @@ JSON.parse = function () {
     removeAdSlotRenderer(searchSectionListRenderer);
   }
 
-  if (r?.entries) {
-    r.entries = r?.entries?.filter(
+  if (r?.entries instanceof Array) {
+    r.entries = r.entries.filter(
       (elm) => !elm?.command?.reelWatchEndpoint?.adClientParams?.isAd
     );
   }